Add cloude2e replacetoken test.
